# deploy/helm/fuzekeys/files/openapi.yaml
...
openapi: 3.1.0
info:
  title: FuzeKeys API
  description: 'FuzeKeys — identity and credential vault. Managed digital identities, per-site accounts,
    encrypted vault assets, capability-token secret brokering, and automated signup workflows.


    DERIVED DOCUMENT: regenerate with `python scripts/export-openapi.py`. Do not hand-edit.'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 2.0.0
paths:
...
  /api/google/signup/{identity_id}:
    post:
      tags:
      - Google Integration
      - Google Integration
      summary: Signup With Identity
      description: 'Create a Google account using the specified identity.


        SECURITY: requires auth and that the identity is owned by the caller.'
      operationId: post_google_signup_by_identity_id
      security:
      - HTTPBearer: []
      parameters:
      - name: identity_id
        in: path
        required: true
        schema:
          type: integer
          title: Identity Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleSignupConfig'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Signup With Identity Api Google Signup  Identity Id  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/google/signup/manual:
    post:
      tags:
      - Google Integration
      - Google Integration
      summary: Manual Signup
      description: 'Create a Google account with manually provided data.


        SECURITY: requires auth. This route drives real signup automation from the

        request body and must not be reachable unauthenticated.'
      operationId: post_google_signup_manual
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualSignupRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Manual Signup Api Google Signup Manual Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/google/config/default:
    get:
      tags:
      - Google Integration
      - Google Integration
      summary: Get Default Config
      description: 'Get the default configuration for Google signup.


        SECURITY: requires auth. Even though it returns no PII, keep the whole

        router authenticated so there is no unauthenticated surface here.'
      operationId: get_google_config_default
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleSignupConfig'
      security:
      - HTTPBearer: []
  /api/google/test/identity-conversion/{identity_id}:
    post:
      tags:
      - Google Integration
      - Google Integration
      summary: Test Identity Conversion
      description: 'Test converting an identity to Google signup data without actually creating an account.


        SECURITY (CRITICAL-1): this endpoint returns DECRYPTED PII (names, username,

        phone, recovery email, birth date, gender). It now requires auth and that the

        identity belongs to the caller; otherwise 404. This closes the unauthenticated

        PII-exfiltration path.'
      operationId: post_google_test_identity_conversion_by_identity_id
      security:
      - HTTPBearer: []
      parameters:
      - name: identity_id
        in: path
        required: true
        schema:
          type: integer
          title: Identity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Test Identity Conversion Api Google Test Identity Conversion  Identity
                  Id  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/google/accounts/{identity_id}:
    get:
      tags:
      - Google Integration
      - Google Integration
      summary: Get Google Accounts
      description: 'Get all Google accounts for a specific identity.


        SECURITY (CRITICAL-1): requires auth; the identity must be owned by the

        caller (404 otherwise). The account lookup is additionally scoped through the

        owned identity so no cross-tenant account data is returned.'
      operationId: get_google_accounts_by_identity_id
      security:
      - HTTPBearer: []
      parameters:
      - name: identity_id
        in: path
        required: true
        schema:
          type: integer
          title: Identity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Google Accounts Api Google Accounts  Identity Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    GoogleSignupConfig:
      properties:
        use_proxy:
          type: boolean
          title: Use Proxy
          default: false
        proxy_config:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Proxy Config
        headless:
          type: boolean
          title: Headless
          default: true
        timeout:
          type: integer
          title: Timeout
          default: 120
        retry_attempts:
          type: integer
          title: Retry Attempts
          default: 3
        use_mobile_user_agent:
          type: boolean
          title: Use Mobile User Agent
          default: false
        custom_user_agent:
          anyOf:
          - type: string
          - type: 'null'
          title: Custom User Agent
        prefer_phone_verification:
          type: boolean
          title: Prefer Phone Verification
          default: true
        auto_handle_captcha:
          type: boolean
          title: Auto Handle Captcha
          default: false
        save_cookies:
          type: boolean
          title: Save Cookies
          default: true
      type: object
      title: GoogleSignupConfig
      description: Configuration for Google signup process.
    GoogleSignupData:
      properties:
        first_name:
          type: string
          maxLength: 50
          minLength: 1
          title: First Name
        last_name:
          type: string
          maxLength: 50
          minLength: 1
          title: Last Name
        username:
          type: string
          maxLength: 30
          minLength: 6
          title: Username
        password:
          type: string
          minLength: 8
          title: Password
        phone_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone Number
          description: Phone number for account recovery
        recovery_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Recovery Email
          description: Recovery email address
        birth_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Birth Date
          description: Date of birth (required for some regions)
        gender:
          anyOf:
          - type: string
          - type: 'null'
          title: Gender
          description: Gender (optional)
        interests:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Interests
          description: User interests
        skip_phone_verification:
          type: boolean
          title: Skip Phone Verification
          description: Try to skip phone verification if possible
          default: false
      type: object
      required:
      - first_name
      - last_name
      - username
      - password
      title: GoogleSignupData
      description: Data structure for Google account signup.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ManualSignupRequest:
      properties:
        signup_data:
          $ref: '#/components/schemas/GoogleSignupData'
        config:
          anyOf:
          - $ref: '#/components/schemas/GoogleSignupConfig'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - signup_data
      title: ManualSignupRequest
      description: 'Explicit body model for POST /signup/manual.


        Previously this endpoint declared two separate body parameters, so FastAPI synthesised

        the wrapper schema `Body_manual_signup_api_google_signup_manual_post` — and a synthesised

        wrapper cannot carry `extra="forbid"`, which is why gate-identifier flagged this route

        (governance/identifier-standard.md 1: a create body must reject fields it does not

        declare, the OWASP API3:2023 mass-assignment guard).


        Declaring the wrapper explicitly is WIRE-COMPATIBLE: two embedded body params already

        produced `{"signup_data": ..., "config": ...}`, and this model produces exactly the same

        JSON, with `config` still optional. What changes is that an undeclared top-level field is

        now a 422 instead of being silently ignored.'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...